The default data type is float and the error bound for verification is 1e-3

+++ slu-cuda/include/type.h
@@ -1,4 +1,4 @@
 #ifndef TYPE_H
 #define TYPE_H
-typedef double REAL;
+typedef float REAL;
 #endif
